Add comments to INPUT (#323)
* Add comment to input_read and input_parse

* add comment to reset_input_read_return

* make seperate files for input_reset_cursor_pos and input_get_curs_row

* add comment to input_reset_cursor-pos

* Add comment to input_get_curs_row

* Add comments to scroll_down_terminal, and input_print_str

* Add comment to input_parse_tab

* Add comment to curs_move_n_left

* Add comments to curs_move_n_right and curs_move_n_left

* Add comment to input_handle_delete

* Add comments to the remaining inputhandling functions that I wrote or could explain well

* Put input_add_chunk in seperate file

* Move input_read_from_buffer into input_read file

* Add comment to input_resize_window
