+++ srcs/input_handling/input_parse_home.c
@@ -6,14 +6,15 @@
 /*   By: omulder <omulder@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/05/16 13:37:33 by rkuijper       #+#    #+#                */
-/*   Updated: 2019/08/30 16:54:27 by jbrinksm      ########   odam.nl         */
+/*   Updated: 2019/10/05 17:37:47 by jbrinksm      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vsh.h"
 
 /*
-**	Calculates steps to home and then uses curs_move_n_left.
+**	Uses the internal `data->line->index` to go to the start of the command
+**	line.
 */
 
 void		curs_go_home(t_vshdata *data)
