+++ srcs/input_handling/input_resize_window_check.c
@@ -6,7 +6,7 @@
 /*   By: jbrinksm <jbrinksm@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/08/29 10:40:21 by jbrinksm       #+#    #+#                */
-/*   Updated: 2019/09/27 12:53:07 by rkuijper      ########   odam.nl         */
+/*   Updated: 2019/10/06 10:35:27 by jbrinksm      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,21 @@
 #include <term.h>
 #include <signal.h>
 #include <sys/ioctl.h>
+
+/*
+**	When the window gets resized, we have to make sure that the complete command
+**	line is properly displayed.
+**
+**	We saved the previous window size and the amount of lines that we
+**	have occupied on the command line. We then use this information to get our
+**	cursor back to the start of the line in this new environment.
+**
+**	When we have done that we will clear the whole line and reprint the prompt
+**	and command line in the new terminal size.
+**
+**	Finally, we reposition the cursor on the character that it was originally
+**	on.
+*/
 
 static void	input_resize_window(t_vshdata *data, struct winsize new)
 {
